set a Mobile's Position and Movement to Doubles to allow for more fine-grained calculations. This might result in a loss of precision, but in a 200 x 200 grid I'm not sure if it will be noticeable

+++ app/game/mobile/MobileModule.scala
@@ -12,8 +12,8 @@ import game.world.RoomModule
 trait MobileModule extends EventModule {
   this: RoomModule ⇒
 
-  case class Movement( val x: Int, val y: Int )
-  case class Position( x: Int, y: Int ) {
+  case class Movement( val x: Double, val y: Double )
+  case class Position( x: Double, y: Double ) {
     lazy val head = ( x, y + 2 )
     lazy val feet = ( x, y - 2 )
     lazy val right = ( x + 1, y )
